fix: stop job immediately when a new job is submitted && add delay option
When a new job is submitted, the previous implementation will wait until
the previous job finishes and reset a new timer, which might be a waste
for llama.cpp computing resources.

The new fix will cancel the previous job immediately to release
resources.

Also, I added a new option `delay` to allow users to customize the
debounce behavior.